src/backend/app/mutil_agent/requirements.txt
fastapi==0.115.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dotenv==1.0.1
httpx==0.27.2
uvicorn==0.32.0
gunicorn==23.0.0
alembic==1.14.0
SQLAlchemy-Utils==0.41.2
langgraph==0.2.74
langchain==0.3.19
langchain-community==0.3.18
pydantic==2.11.7
pydantic_core==2.33.2
boto3==1.36.25
botocore==1.36.25
langsmith
langchain_aws==0.2.13
psycopg2-binary==2.9.10
pillow==11.2.1
beanie==1.29.0
fastapi-pagination==0.12.34
pandas==2.2.0
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
aiohttp==3.10.11
requests==2.32.3
python-multipart==0.0.9
# Lightweight OCR dependencies (much smaller than EasyOCR)
pytesseract==0.3.10
pdf2image==1.17.0
# Note: System packages needed in Dockerfile: tesseract-ocr tesseract-ocr-vie
strands-agents>=0.1.0
strands-agents-tools>=0.1.0